警告削除

+++ Engine/Engine/Audio/Audio.cpp
@@ -208,11 +208,11 @@ int32_t Audio::SoundLoad(const std::filesystem::path& fileName) {
     MFCreateMediaType(&pMFMediaType);
     pMFMediaType->SetGUID(MF_MT_MAJOR_TYPE, MFMediaType_Audio);
     pMFMediaType->SetGUID(MF_MT_SUBTYPE, MFAudioFormat_PCM);
-    pMFSourceReader->SetCurrentMediaType(MF_SOURCE_READER_FIRST_AUDIO_STREAM, nullptr, pMFMediaType);
+    pMFSourceReader->SetCurrentMediaType(DWORD(MF_SOURCE_READER_FIRST_AUDIO_STREAM), nullptr, pMFMediaType);
 
     pMFMediaType->Release();
     pMFMediaType = nullptr;
-    pMFSourceReader->GetCurrentMediaType(MF_SOURCE_READER_FIRST_AUDIO_STREAM, &pMFMediaType);
+    pMFSourceReader->GetCurrentMediaType(DWORD(MF_SOURCE_READER_FIRST_AUDIO_STREAM), &pMFMediaType);
 
     SoundData soundData = {};
     soundData.filename = fileName.string();
@@ -226,7 +226,7 @@ int32_t Audio::SoundLoad(const std::filesystem::path& fileName) {
     while (true) {
         IMFSample* pMFSample{ nullptr };
         DWORD dwStreamFlags{ 0 };
-        pMFSourceReader->ReadSample(MF_SOURCE_READER_FIRST_AUDIO_STREAM, 0, nullptr, &dwStreamFlags, nullptr, &pMFSample);
+        pMFSourceReader->ReadSample(DWORD(MF_SOURCE_READER_FIRST_AUDIO_STREAM), 0, nullptr, &dwStreamFlags, nullptr, &pMFSample);
 
         if (dwStreamFlags & MF_SOURCE_READERF_ENDOFSTREAM) {
             break;
